updated tabs

+++ components/PostsList/PostsList.tsx
@@ -2,10 +2,7 @@ import { useEffect, useMemo, useState } from 'react'
 import styles from './PostsList.module.scss'
 import { deleteContentOutline, getPostsByDomain } from '@/perfect-seo-shared-components/services/services'
 import * as Modal from '@/perfect-seo-shared-components/components/Modal/Modal'
-import moment from 'moment-timezone'
-import useViewport from '@/perfect-seo-shared-components/hooks/useViewport'
 import Loader from '../Loader/Loader'
-import Link from 'next/link'
 import TypeWriterText from '@/perfect-seo-shared-components/components/TypeWriterText/TypeWriterText'
 import PostItem from '../PostItem/PostItem'
 
@@ -21,11 +18,20 @@ const PostsList = ({ domain_name, active }: PostsListProps) => {
 
   const getPosts = () => {
     if (domain_name && active) {
-      getPostsByDomain(domain_name)
-        .then(res => {
-          setData(res.data)
-          setLoading(false)
-        })
+      if (domain_name === 'all') {
+        getPostsByDomain(domain_name)
+          .then(res => {
+            setData(res.data)
+            setLoading(false)
+          })
+      }
+      else {
+        getPostsByDomain(domain_name)
+          .then(res => {
+            setData(res.data)
+            setLoading(false)
+          })
+      }
     }
   }
 
@@ -63,7 +69,7 @@ const PostsList = ({ domain_name, active }: PostsListProps) => {
     let interval;
     if (domain_name && active) {
       getPosts();
-      setInterval(getPosts, 60000)
+      interval = setInterval(getPosts, 60000)
     }
 
     return () => {
@@ -77,45 +83,49 @@ const PostsList = ({ domain_name, active }: PostsListProps) => {
 
   if (!active) return null
 
-  return (
-    <div className={styles.wrap}>
-      <div className='row d-flex justify-content-between align-items-end my-3'>
-        <div className='col-12 col-md-auto d-flex align-items-end'>
-          <h2 className='text-primary mb-0'>
-            <TypeWriterText string="Generated Posts" withBlink />
-          </h2>
-          {filteredData?.length > 0 && <p className='badge rounded-pill text-bg-primary ms-3 d-flex align-items-center mb-1'>{filteredData.length}</p>}
-        </div>
-        <div className='col-auto me-2'>
-          <div className="form-group">
-            <label className="form-label"><strong>Filter</strong></label>
-            <select className="form-control" value={filter} onChange={handleFilterChange}>
-              <option value="all">All</option>
-              <option value="completed">Completed</option>
-              <option value="other">Processing</option>
-            </select>
-          </div>
-        </div>
-      </div>
-      {loading ? <Loader />
-        : filteredData?.length > 0 ?
-          <div className='row d-flex g-3'>
-            {filteredData.map((obj, i) => {
-              return <PostItem post={obj} key={obj.content_plan_outline_guid} refresh={getPosts} />
-            })}</div> :
-          <h5><TypeWriterText withBlink string="The are no results for the given parameters" /></h5>}
-      <Modal.Overlay open={deleteModal} onClose={() => { setDeleteModal(null) }}>
-        <Modal.Title title="Delete Plan" />
-        <Modal.Description>
-          Are you sure you want to delete this post?
-          <div className='d-flex justify-content-between mt-5'>
-            <button onClick={() => { setDeleteModal(null) }} className="btn btn-warning">Cancel</button>
-            <button onClick={(e) => { e.preventDefault(); deleteHandler(deleteModal) }} className="btn btn-primary">Yes</button>
-          </div>
-        </Modal.Description>
-      </Modal.Overlay>
-    </div>
+  if (!domain_name) return (
+    <h5 className='my-5 text-center'><TypeWriterText withBlink string="Please select a domain to view posts" /></h5>
   )
+  else
+    return (
+      <div className={styles.wrap}>
+        <div className='row d-flex justify-content-between align-items-end my-3'>
+          <div className='col-12 col-md-auto d-flex align-items-end'>
+            <h2 className='text-primary mb-0'>
+              <TypeWriterText string="Generated Posts" withBlink />
+            </h2>
+            {filteredData?.length > 0 && <p className='badge rounded-pill text-bg-primary ms-3 d-flex align-items-center mb-1'>{filteredData.length}</p>}
+          </div>
+          <div className='col-auto me-2'>
+            <div className="form-group">
+              <label className="form-label"><strong>Filter</strong></label>
+              <select className="form-control" value={filter} onChange={handleFilterChange}>
+                <option value="all">All</option>
+                <option value="completed">Completed</option>
+                <option value="other">Processing</option>
+              </select>
+            </div>
+          </div>
+        </div>
+        {loading ? <Loader />
+          : filteredData?.length > 0 ?
+            <div className='row d-flex g-3'>
+              {filteredData.map((obj, i) => {
+                return <PostItem post={obj} key={obj.content_plan_outline_guid} refresh={getPosts} />
+              })}</div> :
+            <h5><TypeWriterText withBlink string="The are no results for the given parameters" /></h5>}
+        <Modal.Overlay open={deleteModal} onClose={() => { setDeleteModal(null) }}>
+          <Modal.Title title="Delete Plan" />
+          <Modal.Description>
+            Are you sure you want to delete this post?
+            <div className='d-flex justify-content-between mt-5'>
+              <button onClick={() => { setDeleteModal(null) }} className="btn btn-warning">Cancel</button>
+              <button onClick={(e) => { e.preventDefault(); deleteHandler(deleteModal) }} className="btn btn-primary">Yes</button>
+            </div>
+          </Modal.Description>
+        </Modal.Overlay>
+      </div>
+    )
 
 }
